backend/requirements.txt
asgiref==3.7.2
Django==4.2.9
git+https://github.com/encode/django-rest-framework.git
gunicorn==21.2.0
packaging==23.2
psycopg2-binary==2.9.9
setuptools==69.0.2
sqlparse==0.4.4
typing_extensions==4.9.0
wheel==0.42.0
Pillow==10.2.0
djangorestframework-simplejwt==4.8.0
web3==6.15.1
django-oauth-toolkit
requests_oauthlib
django-extensions
Werkzeug
pyOpenSSL
uuid